requirement.txt
functions-framework==3.*
google-cloud-aiplatform
google-cloud-bigquery
google-auth
pytz